+++ roboteam_ai/src/interface/mainWindow.cpp
@@ -189,7 +189,7 @@ void MainWindow::updateRobotsWidget() {
 
     // reload the widgets completely if a robot is added or removed
     // or if the amount of selected robots is not accurate
-    if (robotsLayout->count() != us.size() || amountOfSelectedRobots != visualizer->getSelectedRobots().size()) {
+    if (robotsLayout->count() != static_cast<int>(us.size()) || amountOfSelectedRobots != static_cast<int>(visualizer->getSelectedRobots().size())) {
         amountOfSelectedRobots = visualizer->getSelectedRobots().size();
         clearLayout(robotsLayout.get());
 
@@ -204,7 +204,7 @@ void MainWindow::updateRobotsWidget() {
             robotsLayout->addWidget(groupBox, 1);
         }
     } else {
-        for (int i = 0; i < us.size(); i++) {
+        for (int i = 0; i < static_cast<int>(us.size()); i++) {
             if (robotsLayout->itemAt(i)) {
                 auto robotwidget = robotsLayout->itemAt(i)->widget();
                 clearLayout(robotwidget->layout());
